Correct case of "Type" in librarymanager links in examples (#41)
Arduino IDE has a feature where opening a link of the form `http://librarymanager[/<type>[/<topic>][#<search keywords>]`
opens Library Manager with the specified filters and keywords prefilled.

Support for specifying the "Type" filter setting was claimed during the implementation of the feature in Arduino IDE
1.x, but that never actually worked. The text in this path was simply ignored by the IDE (a fact that would not be
obvious from the links used by this library, since the "All" type filter setting is the default anyway).

During the reimplementation of the feature in Arduino IDE 2.x, support for specifying the Type filter setting was
actually implemented correctly. Instead of ignoring this path component of the URL, the IDE requires it to match to a
valid "Type" filter value. If there is no match, opening the link has no effect. This match is case sensitive.

The previous URLs used the type value "all", but there is no "all" value in the "Type" filter. Instead, there is an
"All" value, and that exact case must be used in order for the link to function in Arduino IDE 2.x.